refactor(schemas): migrate Phase 13 endpoint schemas — 4 files, 20 classes (#6042)
- prometheus_mcp.py: migrate MCPTool→PrometheusMCPTool, QueryMetricRequest,
  QueryRangeRequest, GetVMMetricsRequest, ListMetricsRequest → schemas_system.py
- multimodal.py: migrate CrossModalSearchRequest, TextProcessingRequest,
  EmbeddingRequest, MultiModalResponse, CrossModalSearchResponse → schemas_knowledge.py
- natural_language_search.py: migrate NLSearchRequest, ParsedQueryResponse,
  SearchResultWithExplanation, NLSearchResponse, QuerySuggestionResponse → schemas_knowledge.py
- knowledge.py: migrate AddFactsRequest, AddUrlRequest, AudioIngestRequest,
  DocsBrowseRequest, OrgKnowledgeConfigPayload → schemas_knowledge.py
  (preserves field_validator @ classmethod validators)

Renames to avoid generic-name collisions:
- MCPTool → PrometheusMCPTool

Cleanup: removed dead-code request class imports in prometheus_mcp.py
(QueryMetricRequest et al. were defined but never used in endpoint logic).
